Ask students to generate core files in /cores
http://smartos.org/bugview/OS-3042 causes coreadm to stop working when
setting a relative path for core files patterns. In order for coreadm to
work again, the instance has to be deleted, a reboot won't do.

It is not clear if this issue is going to be fixed, so instead we ask
students to generate their core files in /cores to minimize the chance
of them setting relative path pattern and triggering the coreadm bug.