Check for the correct minimum value of bar. Replace occurence of max with min.
The instructions ask for the user to find the minimum value of the `bar`
property. The `bar` property of the `Foo` objects created has a staticly
defined value of `4242`. When checking the solution, the value passed in
is zero-fill right shifted and compared with `9` when it should be compared
with `4242`.

In addition, the hint text when the check fails suggests finding max-value
which is in conflict with the instructions which says min-value.