fix print_all_objects_with_a_given_type
Check for the correct minimum value of bar and replace occurence of max
with min.

The instructions ask for the user to find the minimum value of the `bar`
property. The `bar` property of the `Foo` objects created has a staticly
defined value of `4242`. When checking the solution, the value passed in
is zero-fill right shifted and compared with `9` when it should be
compared with `4242`.

In addition, the hint text when the check fails suggests finding
max-value which is in conflict with the instructions which says
min-value.

PR: #15
PR-URL: https://github.com/joyent/node-debug-school/pull/15
Reviewed-By: Julien Gilli <jgilli@fastmail.fm>